Extract prompts to separate module
